Fix the bug where entire lines can't be selected (to be marked)
The problem was that when you select an entire line, the selection text
starts/ends with `\n` and the anchor/focus nodes had siblings of the
<pre> code line as their parent elements, and so the conditions for
"is this a seleciton in a line of code" weren't met.

The fix was to trim leading/trailing whitespace, and to check both
anchor _and_ focus nodes for the code's class name, and also to handle
ranges with multiple rects better.