Fix popover buttons in non-WebKit based browsers
The marker + arrow popovers rely on focus to show/disappear.
This is determined using the `autofocus` prop. Without this, whenever
you click outside of a popover it would stay there.

When you click the button inside the popover (the remove button) the
focus shifts away from the popover, and onBlur we tell the popover to
disappear.

In non-WebKit based browsers here's the order of operations:

1. mousedown (button inside the popover)
2. blur (popover)
3. focus (button)
4. click (button)

Whereas in WebKit, the blur/focus doesn't happen and focus remains on
the popover.

Since the elements are removed from the DOM in the onBlur event (in
step 2), the click handler (step 4) doesn't happen because by that point
there is no element to trigger that event.

The solution is to avoid blurring the popover, which is done with a
`preventDefault` on the mousedown event (step 1).