Fix selection jumping when there's an arrow/marker over the mouse
In terms of DOM structure the svg is after the code, so when you select
text and the mouse goes over an arrow or marker, the selection end is
set to that element. Effectively this means that the selection covers
the entire code (all the elements in the DOM up to the svg).

In order to fix this we need to set `pointer-events: none` on the
arrows.