Add the ability to update marker/arrow colors after they've been added
This helped uncover a bug: there's a duplication of state between
markers and the `fromMarker` and `toMarker` fields in arrows.
This means that those fields should become ids and the `markers` field
in the state should become a map rather than an array (and the same for
the `arrows` field).