Refactor the annotations' DOM structure
Previously the code (<Code>), svg (<Svg>), and selection popover
(<SelectionPopover>) were siblings inside a container div.

Now the svg and popover are siblings in <CodeAnnotations>, which is
itself a sibling of the code content in <Code>.

<Code> is a grid, with a separate column for line annotations, line
numbers and line content. The svg needed to be a sibling of those so
that it could also be positioned in the grid structure (specifically in
the third column, along with the code content). This is so that changes
in sizing of the other two columns (line annotations + numbers) doesn't
affect the positioning of annotations in the svg.

The popover was moved there basically because it was easier to do and
involved fewer (or rather zero) changes to relative position calculation
from pointer events.