Update TypeScript to version 4.4
And also update the clipboard workaround types:

TS 4.4 added the clipboard API to its built in web types. This means
that we can remove our `clipboard.d.ts` workaround.
HOWEVER, Their policy is that when two browser engines (out of the
three most common ones: WebKit, Blink, and Gecko) implement an API
it's considered widespread enough to include in the library.

The problem here is that the third one (Blink) didn't implement the
API according to the spec so we need to go against the type system.