Better nearest point detection for new arrows
When you start drawing an arrow from an existing arrow there used to be
a little gap. This is because the click target for arrows includes some
padding around the arrow, so events near the arrow counted as events
on the arrow. When you started an arrow from that padding you got a gap.

Now, with the help of some geometry, the arrow will always start from
a point on the arrow that's closest to the mouse. Rejoice.

This was done by calling `pointOnLineNearPoint` on all segments of the
existing arrow and picking the closest result.

This also helped me find a bug with my `minBy` function, which was
very silly and had to do with index math.