Use a context for arrow drawing
Right now it's used only in CodeAnnotations but I want to modify that
in a future commit. Right now there's a lot of prop drilling and I'd
rather avoid that.

I'm splitting the state into two contexts:

* `CurrentArrowContext` for the current arrow
* `EventHandlersContext` for the event handlers that need to be set on
  the svg, markers, and arrows.

I'm doing it that way because it's more efficient: the current arrow
will update whenever the mouse moves, but the event handlers and the
components that use them don't need to rerender as often, so I want to
separate them.
This is also why I don't want to put this in a redux store.

Side note: right now the event handlers depend on currentArrow so they
still cause a rerender, but that will change soon.