Completely change arrow drawing and popover interactions
In order to make arrow drawing easier and more capable we decided to
change the mouse interactions with markers and arrows:

Up until now you had to click on a marker/arrow to open its popover,
where you could change its color or delete it. If you started dragging
from the marker/arrow, you would start a new arrow that was completed
once you let go of the mouse over a different marker.

There are several problems with this drag interaction:
1. It's hard to operate using a trackpad
2. It's pretty much impossible to draw arrows that span multiple
   viewports, because you can't scroll mid-drag
3. It made it very hard to add jointed arrows (straight arrows with
   corners) in a way that was consistent with freehand arrow drawing.

We've decided to change the way this works:

* First of all, the popover now opens on _right_ mouse click and
  replaces the default context menu for markers and arrows.
  The rationale for this is that this operation is less common and it
  felt like a reasonable place to "hide" these operations in
  (delete + change color), since they're probably relatively uncommon.
* Arrow drawing is triggered on click, instead of on drag
* An arrow is added once the destination marker is clicked
* When the "Use straight arrows" setting is active, clicking outside of
  a marker while drawing adds a joint.
* You can hit the escape key to cancel the current arrow drawing.

We felt that this "click to add an arrow" interaction solves the three
issues:

1. It's easier to operate because you don't need to keep the mouse
   button down
2. Arrows can span multiple viewports because you can now scroll while
   drawing
3. Jointed arrows make sense: instead of clicking a destination marker,
   just click on the background to add a joint