Change the copy on the home page and add permalink instructions
People don't read text so hopefully showing the same information twice
will help.

I've made the github url parsing code more explicit (it parses the url
in stages now, instead of using an unclear regex pattern).

This allows us to show a more specific error message when the url you've
entered is a github link but not a permalink.